+++ src/main/java/me/srrapero720/embeddiumplus/foundation/borderless/VideoModeHandler.java
@@ -1,0 +1,6 @@
+package me.srrapero720.embeddiumplus.foundation.borderless;
+
+public interface VideoModeHandler  {
+    void changeVideoMode();
+    void setFullscreenMode(boolean mode);
+}