+++ src/main/java/me/srrapero720/embeddiumplus/mixins/impl/borderless/BorderlessWindowMixin.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
 
 @Mixin(Window.class)
-public class WindowMixin {
+public class BorderlessWindowMixin {
     @Redirect(method = "setMode", at = @At(value = "INVOKE", remap = false, target = "Lorg/lwjgl/glfw/GLFW;glfwSetWindowMonitor(JJIIIII)V"))
     private void redirect$glfwSetWindowMonitor(long window, long monitor, int xpos, int ypos, int width, int height, int refreshRate) {
         if (!EmbyConfig.isLoaded()) EmbyConfig.load();